Adds support for external process-sourced credentials
AWS offer the option to source credentials from an external process, which is handy for using things like 1Password to hold your credentials.

This adds support to `getCreds` so these can be read, via a call to `execSync`. Technically this is a bit bad in that we're passing a random string to `exec`, but to make this work this isn't really much option and AWS mention this themselves in their documentation, so "with great power" etc. etc.

https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-sourcing-external.html