Do not use module augmentation in the module’s declaration
Fixes errors such as “Exports and export assignments are not permitted
in module augmentations.” The main declaration file for the package
should not have to declare augmentations of itself. Just define the
export at the root level.

To still support access to the AwsLiteClient interface we can add it to
a namespace of what we are exporting.