fix(types): Replace `stream` with `Readable` in method signatures, do not rely on non-existent `Upload` `@aws-sdk` response type (#195)
* add type tests with `tsd` library

* do not skip lib check

* structure type tests in relevant plugin types package

* axe type test readme

* run npm i in plugin dirs in ci

* in plugin types dir!

* use pushd/popd instead

* Do not use module augmentation in the module’s declaration

Fixes errors such as “Exports and export assignments are not permitted
in module augmentations.” The main declaration file for the package
should not have to declare augmentations of itself. Just define the
export at the root level.

To still support access to the AwsLiteClient interface we can add it to
a namespace of what we are exporting.

There is no primitive type called `stream`. It looks like the
implementation (e.g. put-object.mjs) use an import from node:stream.
Make sure the type does the same.

* regen plugin types

* fix(s3/types): Add ability for methods to define a custom response type.

* test: ensure we can import core aws-lite client types. fix an issue where return type of getAllResponses returned requests, not responses.

* changelog update

---------

Co-authored-by: Martijn van der Ven <martijn@vanderven.se>